Correct SQL on fppta_last_membership_payment sumfield.
It should only include payments (whether soft_credit or not) if they
are tied to a membership for the attributed contact. E.g. the contribution
should not be attributed to the donor unless the donor was paying
for their own membership; soft_credit payments for another member
should only be counted for that member, not the donor.